requirements.txt
requests==2.26.0
docker==4.4.3
pyyaml==5.4.1
websocket-client==0.59.0
